Pivot

+++ database/migrations/2023_04_30_173019_add_type_id_to_projects_table.php
@@ -13,15 +13,18 @@ return new class extends Migration
      */
     public function up()
 {
-    Schema::table('projects', function (Blueprint $table) {
-        $table->unsignedBigInteger('type_id')->nullable();
+    if (!Schema::hasColumn('projects', 'type_id')) {
+        Schema::table('projects', function (Blueprint $table) {
+            $table->unsignedBigInteger('type_id')->nullable();
 
-        $table->foreign('type_id')
-              ->references('id')
-              ->on('types')
-              ->onDelete('cascade');
-    });
+            $table->foreign('type_id')
+                  ->references('id')
+                  ->on('types')
+                  ->onDelete('cascade');
+        });
+    }
 }
+
 
     /**
      * Reverse the migrations.
